microservice-clone-and-process/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
gitpython
torch
transformers
numpy
faiss-cpu
python-magic
tiktoken
openai
pydantic
tenacity